requirements.txt
asn1crypto==0.24.0
Babel==2.5.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.2.2
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-appconf==1.0.2
django-bootstrap-datepicker==1.2.3
django-bootstrap-dynamic-formsets==0.5.0
django-bootstrap-form==3.3
django-bootstrap3==9.1.0
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-formset-js==0.5.0
django-jquery==3.1.0
django-jquery-js==3.1.1
django-phonenumber-field==2.0.0
django-rest-auth==0.9.3
django-rest-knox==3.1.3
django-user-accounts==2.0.3
django-versatileimagefield==1.9
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
idna==2.6
jsonfield==2.0.2
oauthlib==2.0.7
phonenumberslite==8.9.1
Pillow==5.0.0
pinax-cli==1.1.1
pinax-eventlog==2.0.3
pinax-templates==1.0.4
pinax-theme-bootstrap==8.0.1
pinax-webanalytics==4.0.2
pycparser==2.18
PyJWT==1.6.1
pyOpenSSL==17.5.0
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
suit==2.0.2
urllib3==1.22
